Such Work, Much wow.. admin check left till done

+++ view/Admin/Users/Delete.view.php
@@ -2,11 +2,18 @@
 
 if(Router::GetParamByName('ID') !== NULL)
 {
+    if($_SESSION['USER']['USERID'] !== Router::GetParamByName('ID'))
+    {
         $mediaId = Users::DeleteUser(Router::GetParamByName('ID'));
         if(Media::UnlinkImage($mediaId->avatar, true) == true)
         {
             Router::Redirect('/Admin/Users');
         }
+    } 
+    else 
+    {
+        Router::Redirect('/Admin/Users');
+    }
 } 
 else 
 {
