+++ view/Admin/Users/Users.view.php
@@ -1,5 +1,4 @@
 <a href="<?= Router::$BASE . 'Admin/Users/Create' ?>" class="mdl-button mdl-js-button mdl-button--raised mdl-js-ripple-effect mdl-button--accent">Opret Bruger</a>
-<h3>Bruger liste</h3>
 <table class="mdl-data-table mdl-js-data-table mdl-shadow--2dp">
   <thead>
     <tr>
@@ -24,8 +23,11 @@
                 echo '<th class="mdl-data-table__cell--non-numeric"><img src="'. Router::$BASE .$user->filepath.'/150x150_'.$user->filename.'.'.$user->mime.'"></th>';
                 echo '<th class="mdl-data-table__cell--non-numeric">'.$user->roleName.'</th>';
                 echo '<th>'.$user->userKm.'</th>';
-                echo '<td><a href="'.Router::$BASE.'Admin/Users/Edit/'.$user->userId.'"><i class="material-icons">mode_edit</i></td>';
-                echo '<td><a href="'.Router::$BASE.'Admin/Users/Delete/'.$user->userId.'" class="delete"><i class="material-icons">delete</i></td>';
+                if(Permission::LevelCheck($user->roleLevel))
+                {
+                  echo '<td><a href="'.Router::$BASE.'Admin/Users/Edit/'.$user->userId.'"><i class="material-icons">mode_edit</i></td>';
+                  echo '<td><a href="'.Router::$BASE.'Admin/Users/Delete/'.$user->userId.'" class="delete"><i class="material-icons">delete</i></td>';
+                }
                 echo '<tr>';
 
         }
